requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
bandit==1.7.0
boto3==1.17.105
botocore==1.20.105
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.3
django-binary-database-files==1.0.15
django-cors-headers==3.8.0
django-crontab==0.7.1
django-email-bandit==2.0
django-environ==0.4.5
django-filter==2.4.0
django-google-auth==0.4.0
django-googlecalendar==3.0.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-nose==1.4.7
django-oauth-toolkit==1.5.0
django-pandas==0.6.4
django-slack==5.16.2
django-storages==1.11.1
django-webpack-loader==1.0.0
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
drf-extensions==0.7.1
drf-nested-routers==0.93.4
drf-yasg==1.20.0
ephem==3.7.7.1
FeinCMS==1.20.0
gdata==2.0.18
gitdb==4.0.7
GitPython==3.1.17
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.8.2
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jwcrypto==0.9.1
LunarCalendar==0.0.9
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==5.1.0
mysqlclient==2.0.3
nose==1.3.7
openpyxl==3.0.10
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pbr==5.6.0

# pep8
pep8>=1.7.1
pep8-naming>=0.12.1
Pillow==8.2.0

# pre commit
pre-commit>=2.15.0
pre-commit>=2.15.0
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8

# py code style
pycodestyle>=2.7.0
pycparser==2.20
pydocstyle>=6.1.1
pyflakes>=2.3.1 # via flake8
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
q==2.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
s3transfer==0.4.2

setuptools~=59.5.0
six==1.16.0
slack-bolt==1.11.1
slack-sdk==3.13.0
slackclient==2.9.3
smmap==4.0.0
sqlparse==0.4.1
stevedore==3.3.0
typing-extensions==3.10.0.0
Unidecode==1.3.4
uritemplate==3.0.1
urllib3==1.26.4
utils==1.0.1
uWSGI==2.0.19.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.6.0
